Add cyphernode to the project

+++ stop.sh
@@ -1,5 +1,14 @@
 #!/usr/bin/env bash
 
-# Clean up old docker
-docker stop veto &> /dev/null
-docker rm veto &> /dev/null
+# Fetch running Veto instances
+runningVetoInstances=$(docker ps | grep "veto" | awk '{ print $1 }');
+
+if [[ ! ''$runningVetoInstances == '' ]]; then
+  # Clean up old Docker
+  echo 'Stopping Veto instances'
+  docker stop veto &> /dev/null
+  docker rm veto &> /dev/null
+else
+  echo 'No Veto instance to stop'
+fi
+
